# samples/tokens/swagger.yaml
openapi: 3.0.3
info:
  title: Tokens
  version: "1.0"
  description: |-
    Issue, hold, transfer and redeem tokens. Backed by a distributed ledger,
    preserving privacy using Zero Knowledge Proofs. This API definition fronts 4 different services, which each play a different
    role in the ecosystem. 

    To try it out, let the Issuer issue some funds to one of the users. Watch the balances
    get updated and see the transaction history. Transfer it to another account, or redeem
    the tokens to destroy them.
servers:
  # - url: http://localhost:9500/
  #   description: auditor
  - url: http://localhost:9100/
    description: issuer
  - url: http://localhost:9500/
    description: alice and bob
  - url: http://localhost:9600/
    description: carlos and dan
paths:
...
  /issuer/issue:
    servers:
      - url: http://localhost:9100/
        description: issuer
    summary: Issue tokens to an account
    post:
      tags:
        - issuer
      responses:
        "200":
          $ref: "#/components/responses/IssueSuccess"
        default:
          $ref: "#/components/responses/ErrorResponse"
      requestBody:
        description: Instructions to issue funds to an account
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
      operationId: issue
      summary: Issue tokens of any kind to an account

  # Owner
...
  /owner/accounts/{id}/transfer:
    servers:
      - url: http://localhost:9500/
        description: alice and bob
      - url: http://localhost:9600/
        description: carlos and dan
    summary: Transfer tokens to another account
    post:
      tags:
        - owner
      parameters:
        - $ref: "#/components/parameters/id"
          name: id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
      responses:
        "200":
          $ref: "#/components/responses/TransferSuccess"
        default:
          $ref: "#/components/responses/ErrorResponse"
      operationId: transfer
      summary: Transfer tokens to another account
...
components:
  responses:
...
    TransferSuccess:
      description: Success response
      content:
        application/json:
          schema:
            type: object
            required:
              - message
              - payload
            properties:
              message:
                type: string
              payload:
                type: string
                description: Transaction id
          example:
            message: transferred tokens
            payload: 7c26ed6e5e05f7de81a82691b66b1069d1507d9edf3c7b78ea1fa98557ee57b4
...
    IssueSuccess:
      description: Success or error response
      content:
        application/json:
          schema:
            type: object
            required:
              - message
              - payload
            properties:
              message:
                type: string
              payload:
                type: string
                description: Transaction id
...
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Amount:
      description: The amount to issue, transfer or redeem.
      type: object
      required:
        - code
        - value
      properties:
        code:
          description: the code of the token
          default: EURX
          type: string
        value:
          description: value in base units (usually cents)
          type: integer
          format: uint64
          minimum: 0
      example:
        code: EURX
        value: 100
...
    Counterparty:
      description: The counterparty in a Transfer or Issuance transaction.
      required:
        - node
        - account
      type: object
      properties:
        node:
          description: The node that holds the recipient account
          type: string
        account:
          type: string
      example:
        node: owner1
        account: alice
...
    TransferRequest:
      description: Instructions to issue or transfer tokens to an account
      required:
        - counterparty
        - amount
      type: object
      properties:
        amount:
          $ref: "#/components/schemas/Amount"
        counterparty:
          $ref: "#/components/schemas/Counterparty"
        message:
          description: optional message that will be sent and stored with the transfer transaction
          type: string
...
    Error:
      required:
        - message
        - payload
      type: object
      properties:
        message:
          description: High level error message
          type: string
        payload:
          description: Details about the error
          type: string
      example:
        message: error message
        payload: ""
  parameters:
    id:
      name: id
      schema:
        example: alice
        description: account id as registered at the Certificate Authority
        type: string
      in: path
      required: true
...
  securitySchemes: {}
...
security: []